Issue #2: Explicit Remote mode help text re: button logging
Use explicit LED wording
